test(outbound): consolidate message action fixtures
