+++ src/infra/outbound/message-action-execution.reply-action.test.ts
@@ -1,120 +1,13 @@
 // Covers reply-type plugin actions: outbound text hygiene (citation control
 // markers) and current-source delivery marking for implicit reply routes.
-import { afterEach, describe, expect, it, vi } from "vitest";
-import type { ChannelPlugin } from "../../channels/plugins/types.public.js";
-import type { OpenClawConfig } from "../../config/config.js";
+import { afterEach, describe, expect, it } from "vitest";
 import { setActivePluginRegistry } from "../../plugins/runtime.js";
 import { createTestRegistry } from "../../test-utils/channel-plugins.js";
-import { runMessageAction } from "./message-action-runner.js";
-
-type ChannelActionHandler = NonNullable<NonNullable<ChannelPlugin["actions"]>["handleAction"]>;
-
-const testchatConfig = {
-  channels: {
-    testchat: {
-      enabled: true,
-    },
-  },
-} as OpenClawConfig;
-
-function createReplyActionPlugin(handleAction: ChannelActionHandler): ChannelPlugin {
-  return {
-    id: "testchat",
-    meta: {
-      id: "testchat",
-      label: "Test Chat",
-      selectionLabel: "Test Chat",
-      docsPath: "/channels/testchat",
-      blurb: "Reply action test plugin.",
-    },
-    capabilities: { chatTypes: ["direct"] },
-    config: {
-      listAccountIds: () => ["default"],
-      resolveAccount: () => ({ enabled: true }),
-      isConfigured: () => true,
-    },
-    outbound: {
-      deliveryMode: "direct",
-      sendText: async () => ({ channel: "testchat", messageId: "m-send-1" }),
-    },
-    messaging: {
-      targetResolver: {
-        looksLikeId: () => true,
-      },
-    },
-    actions: {
-      describeMessageTool: () => ({ actions: ["reply", "poll"] }),
-      supportsAction: ({ action }) => action === "reply" || action === "poll",
-      handleAction,
-    },
-  };
-}
-
-function registerReplyPlugin() {
-  const payload = { ok: true, messageId: "m-reply-1", repliedTo: "platform-guid-1" };
-  const handleAction = vi.fn(async () => ({
-    content: [{ type: "text" as const, text: JSON.stringify(payload) }],
-    details: payload,
-  }));
-  setActivePluginRegistry(
-    createTestRegistry([
-      { pluginId: "testchat", source: "test", plugin: createReplyActionPlugin(handleAction) },
-    ]),
-  );
-  return handleAction;
-}
-
-async function runReplyAction(params: {
-  actionParams: Record<string, unknown>;
-  currentMessageId?: string | number;
-}) {
-  const toolContext = {
-    currentChannelProvider: "testchat" as const,
-    currentChannelId: "direct:user-1",
-    ...(params.currentMessageId !== undefined ? { currentMessageId: params.currentMessageId } : {}),
-  };
-  return await runMessageAction({
-    cfg: testchatConfig,
-    action: "reply",
-    params: { channel: "testchat", ...params.actionParams },
-    toolContext,
-    messageActionAuthorization: {
-      requesterAccountId: "default",
-      toolContext,
-    },
-    sessionKey: "agent:main:testchat:direct:user-1",
-    defaultAccountId: "default",
-    sourceReplyDeliveryMode: "message_tool_only",
-    dryRun: false,
-  });
-}
-
-async function runPollAction(params: { to: string }) {
-  const toolContext = {
-    currentChannelProvider: "testchat" as const,
-    currentChannelId: "direct:user-1",
-    currentMessageId: "1783",
-  };
-  return await runMessageAction({
-    cfg: testchatConfig,
-    action: "poll",
-    params: {
-      channel: "testchat",
-      to: params.to,
-      pollQuestion: "Preferred default?",
-      pollOption: ["Tell me right away", "Only important"],
-    },
-    toolContext,
-    messageActionAuthorization: {
-      requesterAccountId: "default",
-      toolContext,
-    },
-    sessionKey: "agent:main:testchat:direct:user-1",
-    defaultAccountId: "default",
-    sourceReplyDeliveryMode: "message_tool_only",
-    dryRun: false,
-  });
-}
+import {
+  registerReplyPlugin,
+  runCurrentConversationPollAction as runPollAction,
+  runReplyAction,
+} from "./message-action-test-fixtures.js";
 
 describe("runMessageAction reply-type plugin actions", () => {
   afterEach(() => {
