+++ src/infra/outbound/message-action-runner.plugin-dispatch.test.ts
@@ -3,24 +3,19 @@
 import { createRequireRecord } from "openclaw/plugin-sdk/test-fixtures";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { jsonResult } from "../../agents/tools/common.js";
-import { dispatchChannelMessageAction } from "../../channels/plugins/message-action-dispatch.js";
-import type {
-  ChannelMessageActionContext,
-  ChannelMessageActionName,
-  ChannelPlugin,
-} from "../../channels/plugins/types.public.js";
+import type { ChannelPlugin } from "../../channels/plugins/types.public.js";
 import type { OpenClawConfig } from "../../config/config.js";
-import {
-  normalizeMessagePresentation,
-  renderMessagePresentationFallbackText,
-} from "../../interactive/payload.js";
-import { extractToolPayload } from "../../plugin-sdk/tool-payload.js";
-import { getActivePluginRegistry, setActivePluginRegistry } from "../../plugins/runtime.js";
+import { setActivePluginRegistry } from "../../plugins/runtime.js";
 import { createTestRegistry } from "../../test-utils/channel-plugins.js";
 import { GATEWAY_CLIENT_MODES, GATEWAY_CLIENT_NAMES } from "../../utils/message-channel.js";
-import { runMessageAction } from "./message-action-runner.js";
-
-type ChannelActionHandler = NonNullable<NonNullable<ChannelPlugin["actions"]>["handleAction"]>;
+import {
+  createAlwaysConfiguredPluginConfig,
+  createGatewayActionPlugin,
+  messageActionRunnerMocks as mocks,
+  resetMessageActionRunnerMocks,
+  runMessageAction,
+  setMessageActionTestPlugin as setTestPlugin,
+} from "./message-action-runner.test-helpers.js";
 
 const requireLabeledRecord = createRequireRecord("record", "expected-label");
 
@@ -39,237 +34,9 @@ function expectRecordFields(
   }
 }
 
-const mocks = vi.hoisted(() => ({
-  resolveOutboundChannelPlugin: vi.fn(),
-  executeSendAction: vi.fn(),
-  executePollAction: vi.fn(),
-  hasCorePresentationDelivery: vi.fn(),
-  materializeMessagePresentationFallback: vi.fn(),
-  callGateway: vi.fn(),
-  callGatewayLeastPrivilege: vi.fn(),
-  isGatewayTransportError: vi.fn(),
-  randomIdempotencyKey: vi.fn(() => "idem-gateway-action"),
-  maybeApplyTtsToPayload: vi.fn(async (params: { payload: unknown }) => params.payload),
-  prepareOutboundMirrorRoute: vi.fn(),
-  beginTerminalSourceReplyDelivery: vi.fn(),
-  cancelTerminalSourceReplyDelivery: vi.fn(),
-  isCurrentSourceReplyActionName: vi.fn(() => false),
-  isDeliveredCurrentSourceReply: vi.fn(() => false),
-  isDeliveredCurrentSourceReplyAction: vi.fn(() => false),
-  reconcileTerminalSourceReplyDelivery: vi.fn(),
-}));
-
-vi.mock("./channel-resolution.js", () => ({
-  normalizeDeliverableOutboundChannel: (value?: string | null) =>
-    typeof value === "string" ? value.trim().toLowerCase() || undefined : undefined,
-  resolveOutboundChannelPlugin: mocks.resolveOutboundChannelPlugin,
-  resetOutboundChannelResolutionStateForTest: vi.fn(),
-}));
-
-vi.mock("./outbound-send-service.js", () => ({
-  executeSendAction: mocks.executeSendAction,
-  executePollAction: mocks.executePollAction,
-  hasCorePresentationDelivery: mocks.hasCorePresentationDelivery,
-  materializeMessagePresentationFallback: mocks.materializeMessagePresentationFallback,
-}));
-
-vi.mock("./message.gateway.runtime.js", () => ({
-  callGateway: mocks.callGateway,
-  callGatewayLeastPrivilege: mocks.callGatewayLeastPrivilege,
-  isGatewayTransportError: mocks.isGatewayTransportError,
-  randomIdempotencyKey: mocks.randomIdempotencyKey,
-}));
-
-vi.mock("./source-reply-mirror.js", () => ({
-  beginTerminalSourceReplyDelivery: mocks.beginTerminalSourceReplyDelivery,
-  cancelTerminalSourceReplyDelivery: mocks.cancelTerminalSourceReplyDelivery,
-  isCurrentSourceReplyActionName: mocks.isCurrentSourceReplyActionName,
-  isDeliveredCurrentSourceReply: mocks.isDeliveredCurrentSourceReply,
-  isDeliveredCurrentSourceReplyAction: mocks.isDeliveredCurrentSourceReplyAction,
-  reconcileTerminalSourceReplyDelivery: mocks.reconcileTerminalSourceReplyDelivery,
-}));
-
-vi.mock("../../tts/tts.runtime.js", () => ({
-  maybeApplyTtsToPayload: mocks.maybeApplyTtsToPayload,
-}));
-
-vi.mock("./outbound-session.js", () => ({
-  ensureOutboundSessionEntry: vi.fn(async () => undefined),
-  resolveOutboundSessionRoute: vi.fn(async () => null),
-}));
-
-vi.mock("../../channels/plugins/bootstrap-registry.js", () => ({
-  getBootstrapChannelPlugin: (id: string) =>
-    id === "actionhub"
-      ? {
-          actions: {
-            messageActionTargetAliases: {
-              pin: { aliases: ["messageId"] },
-              unpin: { aliases: ["messageId"] },
-              "list-pins": { aliases: ["chatId"] },
-            },
-          },
-        }
-      : undefined,
-}));
-
-vi.mock("./message-action-threading.js", async () => {
-  const { createOutboundThreadingMock } =
-    await import("./message-action-threading.test-helpers.js");
-  const threading = createOutboundThreadingMock();
-  mocks.prepareOutboundMirrorRoute.mockImplementation(threading.prepareOutboundMirrorRoute);
-  return {
-    ...threading,
-    prepareOutboundMirrorRoute: mocks.prepareOutboundMirrorRoute,
-  };
-});
-
-function setTestPlugin(plugin: unknown, pluginId: string, origin?: "bundled") {
-  setActivePluginRegistry(
-    createTestRegistry([{ pluginId, source: "test", ...(origin ? { origin } : {}), plugin }]),
-  );
-}
-
-function createAlwaysConfiguredPluginConfig(account: Record<string, unknown> = { enabled: true }) {
-  return {
-    listAccountIds: () => ["default"],
-    resolveAccount: () => account,
-    isConfigured: () => true,
-  };
-}
-
-function createGatewayActionPlugin(params: {
-  pluginId: string;
-  label: string;
-  blurb: string;
-  actions: ChannelMessageActionName[];
-  gatewayActions?: ChannelMessageActionName[];
-  capabilities?: ChannelPlugin["capabilities"];
-  messaging?: ChannelPlugin["messaging"];
-  threading?: ChannelPlugin["threading"];
-  handleAction: ChannelActionHandler;
-}): ChannelPlugin {
-  const actions = new Set(params.actions);
-  const gatewayActions = new Set(params.gatewayActions ?? params.actions);
-  return {
-    id: params.pluginId,
-    meta: {
-      id: params.pluginId,
-      label: params.label,
-      selectionLabel: params.label,
-      docsPath: `/channels/${params.pluginId}`,
-      blurb: params.blurb,
-    },
-    capabilities: params.capabilities ?? { chatTypes: ["direct"] },
-    config: createAlwaysConfiguredPluginConfig(),
-    messaging: params.messaging,
-    threading: params.threading,
-    actions: {
-      describeMessageTool: () => ({ actions: params.actions }),
-      supportsAction: ({ action }) => actions.has(action),
-      resolveExecutionMode: ({ action }) => (gatewayActions.has(action) ? "gateway" : "local"),
-      handleAction: params.handleAction,
-    },
-  };
-}
-
-async function executePluginAction(params: {
-  action: "send" | "poll";
-  ctx: Pick<
-    ChannelMessageActionContext,
-    | "channel"
-    | "cfg"
-    | "params"
-    | "mediaAccess"
-    | "accountId"
-    | "gateway"
-    | "toolContext"
-    | "inboundEventKind"
-  > & {
-    dryRun: boolean;
-    agentId?: string;
-  };
-}) {
-  const handled = await dispatchChannelMessageAction({
-    channel: params.ctx.channel,
-    action: params.action,
-    cfg: params.ctx.cfg,
-    params: params.ctx.params,
-    mediaAccess: params.ctx.mediaAccess,
-    mediaLocalRoots: params.ctx.mediaAccess?.localRoots ?? [],
-    mediaReadFile:
-      typeof params.ctx.mediaAccess?.readFile === "function"
-        ? params.ctx.mediaAccess.readFile
-        : undefined,
-    accountId: params.ctx.accountId ?? undefined,
-    gateway: params.ctx.gateway,
-    toolContext: params.ctx.toolContext,
-    inboundEventKind: params.ctx.inboundEventKind,
-    dryRun: params.ctx.dryRun,
-    agentId: params.ctx.agentId,
-  });
-  if (!handled) {
-    throw new Error(`expected plugin to handle ${params.action}`);
-  }
-  return {
-    handledBy: "plugin" as const,
-    payload: extractToolPayload(handled),
-    toolResult: handled,
-  };
-}
-
 describe("runMessageAction plugin dispatch", () => {
   beforeEach(() => {
-    mocks.resolveOutboundChannelPlugin.mockReset();
-    mocks.resolveOutboundChannelPlugin.mockImplementation(
-      ({ channel }: { channel: string }) =>
-        getActivePluginRegistry()?.channels.find((entry) => entry?.plugin?.id === channel)?.plugin,
-    );
-    mocks.executeSendAction.mockReset();
-    mocks.executeSendAction.mockImplementation(
-      async ({ ctx }: { ctx: Parameters<typeof executePluginAction>[0]["ctx"] }) =>
-        await executePluginAction({ action: "send", ctx }),
-    );
-    mocks.executePollAction.mockReset();
-    mocks.executePollAction.mockImplementation(
-      async ({ ctx }: { ctx: Parameters<typeof executePluginAction>[0]["ctx"] }) =>
-        await executePluginAction({ action: "poll", ctx }),
-    );
-    mocks.hasCorePresentationDelivery.mockReset();
-    mocks.hasCorePresentationDelivery.mockImplementation(
-      (outbound?: { sendPayload?: unknown; sendText?: unknown; sendFormattedText?: unknown }) =>
-        Boolean(outbound?.sendPayload || outbound?.sendText || outbound?.sendFormattedText),
-    );
-    mocks.materializeMessagePresentationFallback.mockReset();
-    mocks.materializeMessagePresentationFallback.mockImplementation(
-      (params: { payload: { presentation?: unknown; text?: string }; text?: string }) => {
-        const presentation = normalizeMessagePresentation(params.payload.presentation);
-        const text = (params.text ?? params.payload.text ?? "").trim();
-        if (!presentation) {
-          return text;
-        }
-        const fallback = renderMessagePresentationFallbackText({ presentation });
-        return !fallback || text.includes(fallback)
-          ? text
-          : [text, fallback].filter(Boolean).join("\n\n");
-      },
-    );
-    mocks.callGateway.mockReset();
-    mocks.callGatewayLeastPrivilege.mockReset();
-    mocks.isGatewayTransportError.mockReset();
-    mocks.isGatewayTransportError.mockImplementation(
-      (value: unknown) =>
-        value instanceof Error && (value as { kind?: unknown }).kind === "timeout",
-    );
-    mocks.randomIdempotencyKey.mockClear();
-    mocks.maybeApplyTtsToPayload.mockReset();
-    mocks.maybeApplyTtsToPayload.mockImplementation(
-      async (params: { payload: unknown }) => params.payload,
-    );
-    mocks.prepareOutboundMirrorRoute.mockClear();
-    mocks.beginTerminalSourceReplyDelivery.mockReset();
-    mocks.cancelTerminalSourceReplyDelivery.mockReset();
-    mocks.reconcileTerminalSourceReplyDelivery.mockReset();
+    resetMessageActionRunnerMocks();
   });
   describe("alias-based plugin action dispatch", () => {
     const handleAction = vi.fn(async ({ params }: { params: Record<string, unknown> }) =>
