+++ src/infra/outbound/message-action-runner.poll.test.ts
@@ -1,167 +1,21 @@
 // Covers message-action poll handling through plugin dispatch and core gateway
 // poll fallback.
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { ChannelPlugin } from "../../channels/plugins/types.public.js";
-import type { OpenClawConfig } from "../../config/config.js";
-import { getActivePluginRegistry, setActivePluginRegistry } from "../../plugins/runtime.js";
-import { createTestRegistry } from "../../test-utils/channel-plugins.js";
-import { runMessageAction } from "./message-action-runner.js";
-
-const mocks = vi.hoisted(() => ({
-  executePollAction: vi.fn(),
-  resolveOutboundChannelPlugin: vi.fn(),
-}));
-
-function firstMockArg(
-  mock: { mock: { calls: readonly unknown[][] } },
-  label: string,
-): Record<string, unknown> {
-  const [call] = mock.mock.calls;
-  if (!call) {
-    throw new Error(`expected ${label} call`);
-  }
-  const [arg] = call;
-  if (typeof arg !== "object" || arg === null || Array.isArray(arg)) {
-    throw new Error(`expected ${label} params to be an object`);
-  }
-  return arg as Record<string, unknown>;
-}
-
-vi.mock("./channel-resolution.js", () => ({
-  normalizeDeliverableOutboundChannel: (value?: string | null) =>
-    typeof value === "string" ? value.trim().toLowerCase() || undefined : undefined,
-  resolveOutboundChannelPlugin: mocks.resolveOutboundChannelPlugin,
-  resetOutboundChannelResolutionStateForTest: vi.fn(),
-}));
-
-vi.mock("./outbound-send-service.js", () => ({
-  executeSendAction: vi.fn(async () => {
-    throw new Error("executeSendAction should not run in poll tests");
-  }),
-  executePollAction: mocks.executePollAction,
-}));
-
-vi.mock("./outbound-session.js", () => ({
-  ensureOutboundSessionEntry: vi.fn(async () => undefined),
-  resolveOutboundSessionRoute: vi.fn(async () => null),
-}));
-
-vi.mock("./message-action-threading.js", async () => {
-  const { createOutboundThreadingMock } =
-    await import("./message-action-threading.test-helpers.js");
-  return createOutboundThreadingMock();
-});
-const pollerConfig = {
-  channels: {
-    poller: {
-      botToken: "poller-test",
-    },
-  },
-} as OpenClawConfig;
-
-const pollerTestPlugin: ChannelPlugin = {
-  id: "poller",
-  meta: {
-    id: "poller",
-    label: "Poller",
-    selectionLabel: "Poller",
-    docsPath: "/channels/poller",
-    blurb: "Poller test plugin.",
-  },
-  capabilities: { chatTypes: ["direct", "group"] },
-  config: {
-    listAccountIds: () => ["default"],
-    resolveAccount: () => ({ botToken: "poller-test" }),
-    isConfigured: () => true,
-  },
-  outbound: {
-    deliveryMode: "gateway",
-    sendPoll: async () => ({
-      messageId: "poll-test",
-    }),
-  },
-  messaging: {
-    targetResolver: {
-      looksLikeId: () => true,
-      resolveTarget: async ({ normalized }) => ({
-        to: normalized,
-        kind: "user",
-        source: "normalized",
-      }),
-    },
-  },
-  threading: {
-    resolveAutoThreadId: ({ toolContext, to, replyToId }) => {
-      if (replyToId) {
-        return undefined;
-      }
-      if (toolContext?.currentChannelId !== to) {
-        return undefined;
-      }
-      return toolContext.currentThreadTs;
-    },
-  },
-};
-
-async function runPollAction(params: {
-  cfg: OpenClawConfig;
-  actionParams: Record<string, unknown>;
-  toolContext?: Record<string, unknown>;
-  inboundEventKind?: "user_request" | "room_event";
-}) {
-  await runMessageAction({
-    cfg: params.cfg,
-    action: "poll",
-    params: params.actionParams as never,
-    toolContext: params.toolContext as never,
-    inboundEventKind: params.inboundEventKind,
-  });
-  const call = firstMockArg(mocks.executePollAction, "executePollAction") as {
-    resolveCorePoll?: () => {
-      durationHours?: number;
-      maxSelections?: number;
-      threadId?: string;
-    };
-    ctx?: {
-      plugin?: ChannelPlugin;
-      inboundEventKind?: string;
-      idempotencyKey?: string;
-      params?: Record<string, unknown>;
-    };
-  };
-  return {
-    ...call.resolveCorePoll?.(),
-    ctx: call.ctx,
-  };
-}
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
+import {
+  clearMessageActionPollMocks,
+  messageActionRunnerMocks as mocks,
+  pollerConfig,
+  resetMessageActionPollMocks,
+  runPollAction,
+} from "./message-action-runner.test-helpers.js";
 
 describe("runMessageAction poll handling", () => {
   beforeEach(() => {
-    setActivePluginRegistry(
-      createTestRegistry([
-        {
-          pluginId: "poller",
-          source: "test",
-          plugin: pollerTestPlugin,
-        },
-      ]),
-    );
-    mocks.resolveOutboundChannelPlugin.mockReset();
-    mocks.resolveOutboundChannelPlugin.mockImplementation(
-      ({ channel }: { channel: string }) =>
-        getActivePluginRegistry()?.channels.find((entry) => entry?.plugin?.id === channel)?.plugin,
-    );
-    mocks.executePollAction.mockReset();
-    mocks.executePollAction.mockImplementation(async (input) => ({
-      handledBy: "core",
-      payload: { ok: true, corePoll: input.resolveCorePoll() },
-      pollResult: { ok: true },
-    }));
+    resetMessageActionPollMocks();
   });
 
   afterEach(() => {
-    setActivePluginRegistry(createTestRegistry([]));
-    mocks.executePollAction.mockReset();
+    clearMessageActionPollMocks();
   });
   it("requires at least two poll options", async () => {
     await expect(
